+++ CHANGELOG.md
@@ -13,6 +13,8 @@ BUG FIXES:
   - guests/redhat: Fix networking issues with CentOS. [GH-3649]
   - guests/windows: Human error if WinRM not in use to configure networks. [GH-3651]
   - providers/docker: Show proper error message when on Linux. [GH-3654]
+  - providers/docker: Proxy VM works properly even if default provider
+      environmental variable set to "docker" [GH-3662]
   - synced\_folders/rsync: Create the directory before syncing.
 
 ## 1.6.0 (May 6, 2014)